+++ framework/model/src/main/java/is/codion/framework/model/EntityColumnConditionFactory.java
@@ -18,8 +18,8 @@
  */
 package is.codion.framework.model;
 
+import is.codion.common.model.condition.ColumnConditions.ColumnConditionFactory;
 import is.codion.common.model.condition.ConditionModel;
-import is.codion.common.model.condition.TableConditions.ColumnConditionFactory;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;