+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityColumnConditionFactory.java
@@ -18,8 +18,8 @@
  */
 package is.codion.swing.framework.model;
 
+import is.codion.common.model.condition.ColumnConditions.ColumnConditionFactory;
 import is.codion.common.model.condition.ConditionModel;
-import is.codion.common.model.condition.TableConditions.ColumnConditionFactory;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.ForeignKey;