+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityTableModel.java
@@ -18,9 +18,9 @@
  */
 package is.codion.swing.framework.model;
 
+import is.codion.common.model.condition.ColumnConditions;
+import is.codion.common.model.condition.ColumnConditions.ColumnConditionFactory;
 import is.codion.common.model.condition.ConditionModel;
-import is.codion.common.model.condition.TableConditions;
-import is.codion.common.model.condition.TableConditions.ColumnConditionFactory;
 import is.codion.common.value.Value;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
@@ -249,7 +249,7 @@ public class SwingEntityTableModel extends AbstractEntityTableModel<SwingEntityE
 	}
 
 	@Override
-	public final TableConditions<Attribute<?>> filters() {
+	public final ColumnConditions<Attribute<?>> filters() {
 		return filterModel().filters();
 	}